remove some unused Winsock interop stuff
