+++ src/System.Net.Sockets/src/System.Net.Sockets.csproj
@@ -222,12 +222,6 @@
     <Compile Include="$(CommonPath)\Interop\Windows\Winsock\Interop.WSAConnect.cs">
       <Link>Interop\Windows\Winsock\Interop.WSAConnect.cs</Link>
     </Compile>
-    <Compile Include="$(CommonPath)\Interop\Windows\Winsock\Interop.WSAEnumNetworkEvents.cs">
-      <Link>Interop\Windows\Winsock\Interop.WSAEnumNetworkEvents.cs</Link>
-    </Compile>
-    <Compile Include="$(CommonPath)\Interop\Windows\Winsock\Interop.WSAEventSelect.cs">
-      <Link>Interop\Windows\Winsock\Interop.WSAEventSelect.cs</Link>
-    </Compile>
     <Compile Include="$(CommonPath)\Interop\Windows\Winsock\Interop.WSAGetOverlappedResult.cs">
       <Link>Interop\Windows\Winsock\Interop.WSAGetOverlappedResult.cs</Link>
     </Compile>
